requirements.txt
Flask
pyTelegramBotAPI
requests
Pillow
pytesseract
